Deleted old mapbox isolines implementation
