+++ server/lib/python/cartodb_services/cartodb_services/metrics/config.py
@@ -224,7 +224,6 @@ class IsolinesRoutingConfig(ServiceConfig):
     GEOCODER_PROVIDER_KEY = 'geocoder_provider'
     MAPZEN_PROVIDER = 'mapzen'
     MAPBOX_PROVIDER = 'mapbox'
-    MAPBOX_ISO_PROVIDER = 'mapbox_iso'
     TOMTOM_PROVIDER = 'tomtom'
     HEREMAPS_PROVIDER = 'heremaps'
     DEFAULT_PROVIDER = MAPBOX_PROVIDER
@@ -256,12 +255,8 @@ class IsolinesRoutingConfig(ServiceConfig):
             self._mapzen_matrix_service_params = db_config.mapzen_matrix_service_params
             self._mapzen_isochrones_service_params = db_config.mapzen_isochrones_service_params
         elif self._isolines_provider == self.MAPBOX_PROVIDER:
-            self._mapbox_matrix_api_keys = self._db_config.mapbox_matrix_api_keys
-            self._mapbox_matrix_service_params = db_config.mapbox_matrix_service_params
-            self._mapbox_isochrones_service_params = db_config.mapbox_isochrones_service_params
-        elif self._isolines_provider == self.MAPBOX_ISO_PROVIDER:
-            self._mapbox_iso_isolines_api_keys = self._db_config.mapbox_iso_isolines_api_keys
-            self._mapbox_iso_isolines_service_params = db_config.mapbox_iso_isolines_service_params
+            self._mapbox_isolinesx_api_keys = self._db_config.mapbox_isolines_api_keys
+            self._mapbox_isolines_service_params = db_config.mapbox_isolines_service_params
         elif self._isolines_provider == self.TOMTOM_PROVIDER:
             self._tomtom_isolinesx_api_keys = self._db_config.tomtom_isolines_api_keys
             self._tomtom_isolines_service_params = db_config.tomtom_isolines_service_params
@@ -274,8 +269,6 @@ class IsolinesRoutingConfig(ServiceConfig):
             return 'mapzen_isolines'
         elif self._isolines_provider == self.MAPBOX_PROVIDER:
             return 'mapbox_isolines'
-        elif self._isolines_provider == self.MAPBOX_ISO_PROVIDER:
-            return 'mapbox_iso_isolines'
         elif self._isolines_provider == self.TOMTOM_PROVIDER:
             return 'tomtom_isolines'
 
@@ -324,32 +317,16 @@ class IsolinesRoutingConfig(ServiceConfig):
         return self._isolines_provider == self.MAPZEN_PROVIDER
 
     @property
-    def mapbox_matrix_api_keys(self):
-        return self._mapbox_matrix_api_keys
+    def mapbox_isolines_api_keys(self):
+        return self._mapbox_isolines_api_keys
 
     @property
-    def mapbox_matrix_service_params(self):
-        return self._mapbox_matrix_service_params
-
-    @property
-    def mapbox_isochrones_service_params(self):
-        return self._mapbox_isochrones_service_params
+    def mapbox_isolines_service_params(self):
+        return self._mapbox_isolines_service_params
 
     @property
     def mapbox_provider(self):
         return self._isolines_provider == self.MAPBOX_PROVIDER
-
-    @property
-    def mapbox_iso_isolines_api_keys(self):
-        return self._mapbox_iso_isolines_api_keys
-
-    @property
-    def mapbox_iso_isolines_service_params(self):
-        return self._mapbox_iso_isolines_service_params
-
-    @property
-    def mapbox_iso_provider(self):
-        return self._isolines_provider == self.MAPBOX_ISO_PROVIDER
 
     @property
     def tomtom_isolines_api_keys(self):
@@ -640,7 +617,6 @@ class ServicesDBConfig:
         self._get_here_config()
         self._get_mapzen_config()
         self._get_mapbox_config()
-        self._get_mapbox_iso_config()
         self._get_tomtom_config()
         self._get_geocodio_config()
         self._get_data_observatory_config()
@@ -695,26 +671,15 @@ class ServicesDBConfig:
             raise ConfigException('Mapbox configuration missing')
 
         mapbox_conf = json.loads(mapbox_conf_json)
-        self._mapbox_matrix_api_keys = mapbox_conf['matrix']['api_keys']
-        self._mapbox_matrix_quota = mapbox_conf['matrix']['monthly_quota']
-        self._mapbox_matrix_service_params = mapbox_conf['matrix'].get('service', {})
-        self._mapbox_isochrones_service_params = mapbox_conf.get('isochrones', {}).get('service', {})
+        self._mapbox_isolines_api_keys = mapbox_conf['isolines']['api_keys']
+        self._mapbox_isolines_quota = mapbox_conf['isolines']['monthly_quota']
+        self._mapbox_isolines_service_params = mapbox_conf.get('isolines', {}).get('service', {})
         self._mapbox_routing_api_keys = mapbox_conf['routing']['api_keys']
         self._mapbox_routing_quota = mapbox_conf['routing']['monthly_quota']
         self._mapbox_routing_service_params = mapbox_conf['routing'].get('service', {})
         self._mapbox_geocoder_api_keys = mapbox_conf['geocoder']['api_keys']
         self._mapbox_geocoder_quota = mapbox_conf['geocoder']['monthly_quota']
         self._mapbox_geocoder_service_params = mapbox_conf['geocoder'].get('service', {})
-
-    def _get_mapbox_iso_config(self):
-        mapbox_iso_conf_json = self._get_conf('mapbox_iso_conf')
-        if not mapbox_iso_conf_json:
-            raise ConfigException('Mapbox True Isochrones configuration missing')
-
-        mapbox_iso_conf = json.loads(mapbox_iso_conf_json)
-        self._mapbox_iso_isolines_api_keys = mapbox_iso_conf['isolines']['api_keys']
-        self._mapbox_iso_isolines_quota = mapbox_iso_conf['isolines']['monthly_quota']
-        self._mapbox_iso_isolines_service_params = mapbox_iso_conf.get('isolines', {}).get('service', {})
 
     def _get_tomtom_config(self):
         tomtom_conf_json = self._get_conf('tomtom_conf')
@@ -836,20 +801,16 @@ class ServicesDBConfig:
         return self._mapzen_geocoder_service_params
 
     @property
-    def mapbox_matrix_api_keys(self):
-        return self._mapbox_matrix_api_keys
+    def mapbox_isolines_api_keys(self):
+        return self._mapbox_isolines_api_keys
 
     @property
-    def mapbox_matrix_monthly_quota(self):
-        return self._mapbox_matrix_quota
+    def mapbox_isolines_monthly_quota(self):
+        return self._mapbox_isolines_quota
 
     @property
-    def mapbox_matrix_service_params(self):
-        return self._mapbox_matrix_service_params
-
-    @property
-    def mapbox_isochrones_service_params(self):
-        return self._mapbox_isochrones_service_params
+    def mapbox_isolines_service_params(self):
+        return self._mapbox_isolines_service_params
 
     @property
     def mapbox_routing_api_keys(self):
@@ -874,18 +835,6 @@ class ServicesDBConfig:
     @property
     def mapbox_geocoder_service_params(self):
         return self._mapbox_geocoder_service_params
-
-    @property
-    def mapbox_iso_isolines_api_keys(self):
-        return self._mapbox_iso_isolines_api_keys
-
-    @property
-    def mapbox_iso_isolines_monthly_quota(self):
-        return self._mapbox_iso_isolines_quota
-
-    @property
-    def mapbox_iso_isolines_service_params(self):
-        return self._mapbox_iso_isolines_service_params
 
     @property
     def tomtom_isolines_api_keys(self):
